Adjustments made. Bug corrected.
Added information to the readme
Changed the field "revenue" from Movie to long from int
MovieDetails wasn't being correctly loaded after clicking on one of the buttons to load an alternate list. Bug was corrected by changing the code to another componenent (where it should've been to start with)

16 more movies added, including 10 from 2017 in order to test the top revenue per year list.

+++ MovieListServer/src/main/java/com/movieList/Service/MovieDTO.java
@@ -1,6 +1,6 @@
 package com.movieList.Service;
 
-public record MovieDTO(long id, String movieTitle, int releaseYear, String description, String director, String[] actors, int runtime, double rating, int votes, int revenue, int metascore) {
+public record MovieDTO(long id, String movieTitle, int releaseYear, String description, String director, String[] actors, int runtime, double rating, int votes, long revenue, int metascore) {
 
 }
